+++ StyleFrame/command_line/commandline.py
@@ -6,6 +6,8 @@ import pandas as pd
 import sys
 
 from collections import defaultdict
+from pprint import pprint
+
 from .. import StyleFrame, Container, Styler, version
 from .tests.json_schema import commandline_json_schema
 
@@ -28,8 +30,8 @@ class CommandLineInterface(object):
     def parse_as_json(self):
         try:
             self._load_from_json()
-        except TypeError as ex:
-            print('Got the following error:\n{}\nExiting.'.format(ex))
+        except (TypeError, ValueError) as ex:
+            print('Got the following error:\n{}.'.format(ex))
             return
         self._save()
 
@@ -42,7 +44,10 @@ class CommandLineInterface(object):
         else:
             raise TypeError('Neither --json nor --json_path were provided.')
 
-        jsonschema.validate(sheets, commandline_json_schema)
+        try:
+            jsonschema.validate(sheets, commandline_json_schema)
+        except jsonschema.ValidationError as validation_error:
+            raise ValueError(validation_error)
 
         for sheet in sheets:
             self._load_sheet(sheet)
@@ -95,12 +100,14 @@ def get_cli_args():
                        help='print versions of the Python interpreter, openpyxl, pandas and StyleFrame then quit')
     group.add_argument('--json_path', help='path to json file which defines the Excel file')
     group.add_argument('--json', help='json string which defines the Excel file')
+    group.add_argument('--show-schema', action='store_true', help='Print the JSON schema used for validation and exit',
+                       default=False)
     parser.add_argument('--output_path', help='path of output Excel file, defaults to output.xlsx',
                         default='output.xlsx')
 
     cli_args = parser.parse_args()
 
-    if not cli_args.version and not any((cli_args.json_path, cli_args.json)):
+    if not any((cli_args.version, cli_args.show_schema)) and not any((cli_args.json_path, cli_args.json)):
         parser.error('Either --json_path or --json are required when not using -v.')
 
     return cli_args
@@ -110,6 +117,9 @@ def execute_from_command_line():
     cli_args = get_cli_args()
     if cli_args.version:
         print(version.get_all_versions())
+        return
+    if cli_args.show_schema:
+        pprint(commandline_json_schema)
         return
     CommandLineInterface(input_path=cli_args.json_path, input_json=cli_args.json,
                          output_path=cli_args.output_path).parse_as_json()